refactor: Sketch 추상화
