+++ app/src/main/java/woowacourse/paint/model/MyColor.kt
@@ -1,4 +1,4 @@
-package woowacourse.paint.util
+package woowacourse.paint.model
 
 import android.graphics.Color
