+++ app/src/main/java/woowacourse/paint/model/Sketch.kt
@@ -1,0 +1,7 @@
+package woowacourse.paint.model
+
+import android.graphics.Canvas
+
+abstract class Sketch {
+    abstract fun draw(canvas: Canvas)
+}
